getting back to useEffect
