+++ packages/strapi-design-system/src/Select/SelectButton.js
@@ -2,6 +2,7 @@ import React, { forwardRef } from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import { KeyboardKeys } from '../helpers/keyboardKeys';
+import { DownState, UpState } from './constants';
 
 const StyledSelectButton = styled.button`
   text-align: left;
@@ -29,13 +30,13 @@ export const SelectButton = forwardRef(({ children, labelledBy, expanded, onTrig
       case KeyboardKeys.SPACE:
       case KeyboardKeys.ENTER: {
         e.preventDefault();
-        onTrigger('down');
+        onTrigger(DownState.Keyboard);
         break;
       }
 
       case KeyboardKeys.UP: {
         e.preventDefault();
-        onTrigger('up');
+        onTrigger(UpState.Keyboard);
         break;
       }
