+++ packages/strapi-design-system/src/Select/constants.js
@@ -1,0 +1,6 @@
+export const DownState = {
+  Keyboard: 'down:keyboard',
+  Mouse: 'down:mouse',
+};
+
+export const UpState = { Keyboard: 'up:keyboard', Mouse: 'up:mouse' };
