+++ packages/strapi-design-system/src/Select/hooks/useListRef.js
@@ -1,0 +1,37 @@
+import { useRef, useEffect } from 'react';
+import { DownState, UpState } from '../constants';
+import { changeDescendant } from '../utils';
+
+export const useListRef = (expanded, onSelectItem, multi) => {
+  const listRef = useRef(null);
+
+  useEffect(() => {
+    listRef.current.focus();
+  }, []);
+
+  useEffect(() => {
+    if (!listRef.current) return;
+    const lastSelected = listRef.current.querySelector('[aria-selected="true"]');
+    const options = listRef.current.querySelectorAll('[role="option"]');
+
+    let nextOption;
+
+    if (lastSelected) {
+      nextOption = lastSelected;
+    } else if (expanded === UpState.Keyboard) {
+      nextOption = options[options.length - 1];
+    } else if (expanded === DownState.Keyboard) {
+      nextOption = options[0];
+    }
+
+    if (nextOption) {
+      changeDescendant(listRef.current, nextOption);
+
+      if (!multi) {
+        onSelectItem(nextOption.getAttribute('data-strapi-value'));
+      }
+    }
+  }, []);
+
+  return listRef;
+};
